refactor(ui): move non-component exports to sibling modules (fast-refresh)
Part of #418 (react-refresh/only-export-components, 9 -> 0). Each component file
now exports only its component; shared constants/helpers move to siblings:
- RecentChangesSection -> RecentChangesSection.helpers.jsx (OP_STYLES,
  KIND_TO_TAB, renderSummaryWithLink); EntityTimeline imports from there.
- RiskScoreSection -> RiskScoreSection.constants.js (RISK_FIELDS); Resource/User
  detail pages updated.
- ContextPicker -> ContextPicker.helpers.js (matchesTargetTypes).
- ContextTreeView -> ContextTreeView.helpers.js (dedupeSegments, parseOrg,
  computeChildLabels, stripSiblingPrefix + private deps).
Co-located unit tests re-point to the helper modules. No behaviour change.
Verified: 642 UI tests pass (2 failures are the pre-existing #495 locale bug).
